# schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Setmark Home Server",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/sync/pull": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Pull",
        "operationId": "sync_pull_api_sync_pull_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return rows updated after this timestamp. Omit for full pull.",
              "title": "Since"
            },
            "description": "Return rows updated after this timestamp. Omit for full pull."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncPullOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/results": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Results",
        "description": "App pushes completed workout data. Idempotent \u2014 UUIDs prevent duplicates.\n\nEvery set_log and status update is scoped to the authenticated user. Set logs\nfor workout_items belonging to another user are rejected; status updates for\nanother user's workouts 404.",
        "operationId": "sync_results_api_sync_results_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncResultsIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResultsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityIntentIn": {
        "properties": {
          "activity_domain": {
            "type": "string",
            "enum": [
              "running",
              "cycling",
              "rowing",
              "swimming",
              "walking",
              "hiking",
              "functional_strength",
              "traditional_strength",
              "hiit",
              "mobility",
              "mixed_modal",
              "carry",
              "other"
            ],
            "title": "Activity Domain"
          },
          "environment": {
            "type": "string",
            "enum": [
              "indoor",
              "outdoor",
              "unspecified"
            ],
            "title": "Environment",
            "default": "unspecified"
          },
          "preservation_policy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "preserve_primary_activity",
                  "preserve_structure",
                  "preserve_elapsed_time",
                  "preserve_distance",
                  "preserve_mixed_modality"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Preservation Policy"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "activity_domain"
        ],
        "title": "ActivityIntentIn"
      },
      "ExerciseLastPerformed": {
        "properties": {
          "exercise_id": {
            "type": "string",
            "title": "Exercise Id"
          },
          "last_set_logs": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveSetLogRead"
            },
            "type": "array",
            "title": "Last Set Logs"
          }
        },
        "type": "object",
        "required": [
          "exercise_id",
          "last_set_logs"
        ],
        "title": "ExerciseLastPerformed",
        "description": "Inline history the app uses to show 'last time you did this exercise.'"
      },
      "ExerciseRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "demo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Url"
          },
          "default_prescription_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Prescription Json"
          },
          "default_alternatives_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Alternatives Json"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "notes",
          "demo_url"
        ],
        "title": "ExerciseRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrimitiveBlockIn": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "repeat": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Repeat",
            "default": 1
          },
          "work_target": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveWorkTargetIn"
            },
            "type": "array",
            "title": "Work Target"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveSetIn"
            },
            "type": "array",
            "title": "Sets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "sets"
        ],
        "title": "PrimitiveBlockIn"
      },
...
      "PrimitiveSetLogIn": {
        "oneOf": [
          {
            "properties": {
              "role": {
                "const": "slot"
              }
            },
            "required": [
              "slot_id",
              "set_id",
              "block_id",
              "set_index"
            ]
          },
          {
            "not": {
              "anyOf": [
                {
                  "required": [
                    "slot_id"
                  ]
                },
                {
                  "required": [
                    "planned_exercise_id"
                  ]
                },
                {
                  "required": [
                    "performed_exercise_id"
                  ]
                }
              ]
            },
            "properties": {
              "role": {
                "const": "set_result"
              },
              "set_index": {
                "const": 0
              }
            },
            "required": [
              "role",
              "set_id",
              "block_id"
            ]
          },
          {
            "not": {
              "anyOf": [
                {
                  "required": [
                    "slot_id"
                  ]
                },
                {
                  "required": [
                    "set_id"
                  ]
                },
                {
                  "required": [
                    "planned_exercise_id"
                  ]
                },
                {
                  "required": [
                    "performed_exercise_id"
                  ]
                }
              ]
            },
            "properties": {
              "role": {
                "const": "block_result"
              },
              "set_index": {
                "const": 0
              },
              "set_repeat_index": {
                "const": 0
              }
            },
            "required": [
              "role",
              "block_id"
            ]
          }
        ],
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "slot",
              "set_result",
              "block_result"
            ],
            "title": "Role",
            "default": "slot"
          },
          "slot_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Id"
          },
          "set_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Set Id"
          },
          "block_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Id"
          },
          "workout_id": {
            "type": "string",
            "title": "Workout Id"
          },
          "planned_exercise_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Exercise Id"
          },
          "performed_exercise_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Performed Exercise Id"
          },
          "set_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Set Index",
            "default": 0
          },
          "set_repeat_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Set Repeat Index",
            "default": 0
          },
          "block_repeat_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Block Repeat Index",
            "default": 0
          },
          "reps": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight"
          },
          "weight_unit": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "kg",
                  "lb"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Unit"
          },
          "duration_sec": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Sec"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "rounds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rounds"
          },
          "rir": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rir"
          },
          "hr_avg_bpm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hr Avg Bpm"
          },
          "hr_max_bpm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hr Max Bpm"
          },
          "is_warmup": {
            "type": "boolean",
            "title": "Is Warmup",
            "default": false
          },
          "skipped": {
            "type": "boolean",
            "title": "Skipped",
            "default": false
          },
          "side": {
            "type": "string",
            "enum": [
              "left",
              "right",
              "bilateral"
            ],
            "title": "Side",
            "default": "bilateral"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Completed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "workout_id",
          "completed_at"
        ],
        "title": "PrimitiveSetLogIn",
        "description": "Primitive result row. The role selects the coordinate grammar: slot rows identify block/set/slot and must include set_index as the commit sequence within the block/set repeat instance; set_result rows identify block/set, use set_index 0 as their aggregate sentinel, and cannot carry slot or exercise ids; block_result rows identify block, use set_index and set_repeat_index 0 as aggregate sentinels, and cannot carry set, slot, or exercise ids. Runtime validation also verifies the coordinates against the persisted primitive workout tree."
      },
...
      "SyncPullOut": {
        "properties": {
          "workouts": {
            "items": {
              "$ref": "#/components/schemas/WorkoutRead"
            },
            "type": "array",
            "title": "Workouts"
          },
          "exercises": {
            "items": {
              "$ref": "#/components/schemas/ExerciseRead"
            },
            "type": "array",
            "title": "Exercises"
          },
          "user_parameters": {
            "items": {
              "$ref": "#/components/schemas/UserParameterRead"
            },
            "type": "array",
            "title": "User Parameters"
          },
          "last_performed": {
            "items": {
              "$ref": "#/components/schemas/ExerciseLastPerformed"
            },
            "type": "array",
            "title": "Last Performed"
          },
          "server_time": {
            "type": "string",
            "title": "Server Time"
          }
        },
        "type": "object",
        "required": [
          "workouts",
          "exercises",
          "user_parameters",
          "last_performed",
          "server_time"
        ],
        "title": "SyncPullOut"
      },
      "SyncResultsIn": {
        "properties": {
          "primitive_set_logs": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveSetLogIn"
            },
            "type": "array",
            "title": "Primitive Set Logs"
          },
          "status_updates": {
            "items": {
              "$ref": "#/components/schemas/WorkoutStatusUpdate"
            },
            "type": "array",
            "title": "Status Updates"
          },
          "workout_resets": {
            "items": {
              "$ref": "#/components/schemas/WorkoutReset"
            },
            "type": "array",
            "title": "Workout Resets"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SyncResultsIn",
        "description": "App pushes this when a workout finishes (or on next connectivity)."
      },
      "SyncResultsOut": {
        "properties": {
          "primitive_set_logs_received": {
            "type": "integer",
            "title": "Primitive Set Logs Received"
          },
          "status_updates_received": {
            "type": "integer",
            "title": "Status Updates Received"
          },
          "workout_resets_received": {
            "type": "integer",
            "title": "Workout Resets Received"
          }
        },
        "type": "object",
        "required": [
          "primitive_set_logs_received",
          "status_updates_received",
          "workout_resets_received"
        ],
        "title": "SyncResultsOut"
      },
...
      "UserParameterRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "key": {
            "type": "string",
            "title": "Key"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "key",
          "value",
          "updated_at",
          "source"
        ],
        "title": "UserParameterRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkoutRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "scheduled_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "tags_json": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags Json"
          },
          "activity_intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActivityIntentIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "primitive_blocks": {
            "items": {
              "$ref": "#/components/schemas/PrimitiveBlockIn"
            },
            "type": "array",
            "minItems": 1,
            "title": "Primitive Blocks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "name",
          "scheduled_date",
          "status",
          "source",
          "notes",
          "tags_json",
          "created_at",
          "updated_at",
          "completed_at",
          "primitive_blocks"
        ],
        "title": "WorkoutRead"
      },
      "WorkoutReset": {
        "properties": {
          "workout_id": {
            "type": "string",
            "title": "Workout Id"
          }
        },
        "type": "object",
        "required": [
          "workout_id"
        ],
        "title": "WorkoutReset",
        "description": "App tells server to erase same-day execution data for a workout.\n\nThis is the inverse of a completed-workout push: delete the primitive_set_logs\ntied to the workout tree and put the workout back into `planned` so a subsequent\npull does not resurrect a locally reset History row."
      },
      "WorkoutStatusUpdate": {
        "properties": {
          "workout_id": {
            "type": "string",
            "title": "Workout Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "completed",
              "skipped"
            ],
            "title": "Status"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "workout_id",
          "status"
        ],
        "title": "WorkoutStatusUpdate",
        "description": "App tells server 'this workout transitioned to status X at time Y.'\n\n`notes` is optional: the app carries the user's post-workout note as\npart of the terminal status push so the server is authoritative for\nthe value. Without this, the local cache held the note but the next\n`sync/pull` overwrote it with the server's stale value (bug: post-\ncompletion notes disappeared on next pull). Sending `None` leaves\nthe existing server-side `notes` untouched \u2014 only a non-None value\nupdates the row."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}